Correcoes

+++ cuda/host.cu
@@ -58,11 +58,11 @@ __global__ void box_filter_kernel_8u_c1(unsigned char* output,const int width, c
 }
 
 
-float box_filter_8u_c1(initialParams* ct, PPMImageParams* imageParams, PPMBlock* block, int numBlock, cudaStream_t* streamSmooth)
+double box_filter_8u_c1(initialParams* ct, PPMImageParams* imageParams, PPMBlock* block, int numBlock, cudaStream_t* streamSmooth)
 {
 
     cudaEvent_t start, stop;
-    float time;
+    double time;
     cudaEventCreate(&start);
     cudaEventCreate(&stop);
